Reimplement UnreachableListener as MemberRemovedListener
